Separate annotation manager class and clean up
